requirements.txt
numpy
Pillow
PyYAML
trimesh
imageio


